MAINT: Reorganize _digitization
