--- mne/_digitization/__init__.py
@@ -1,5 +1,0 @@
-from .base import DigPoint, Digitization
-
-__all__ = [
-    'DigPoint',
-]
