+++ mne/channels/tests/test_montage.py
@@ -29,10 +29,9 @@ from mne.channels.montage import (_set_montage, transform_to_head,
 from mne.utils import _TempDir, run_tests_if_main, assert_dig_allclose
 from mne.bem import _fit_sphere
 from mne.io.constants import FIFF
-from mne._digitization import Digitization
-from mne._digitization._utils import _format_dig_points
-from mne._digitization._utils import _get_fid_coords
-from mne._digitization.base import _get_dig_eeg, _count_points_by_type
+from mne.io._digitization import (Digitization, _format_dig_points,
+                                  _get_fid_coords, _get_dig_eeg,
+                                  _count_points_by_type)
 
 from mne.viz._3d import _fiducial_coords
