+++ mne/io/kit/kit.py
@@ -29,7 +29,7 @@ from .constants import KIT, LEGACY_AMP_PARAMS
 from .coreg import read_mrk
 from ...event import read_events
 
-from ..._digitization._utils import _set_dig_kit
+from .._digitization import _set_dig_kit
 
 
 def _call_digitization(info, mrk, elp, hsp):
