+++ mne/io/meas_info.py
@@ -29,13 +29,10 @@ from .write import (start_file, end_file, start_block, end_block,
 from .proc_history import _read_proc_history, _write_proc_history
 from ..transforms import invert_transform, Transform
 from ..utils import logger, verbose, warn, object_diff, _validate_type
-from .._digitization.base import _format_dig_points
+from ._digitization import (_format_dig_points, _dig_kind_proper,
+                            _dig_kind_rev, _dig_kind_ints, _read_dig_fif)
+from ._digitization import write_dig as _dig_write_dig
 from .compensator import get_current_comp
-
-# XXX: most probably the functions needing this, should go somewhere else
-from .._digitization.base import _dig_kind_proper, _dig_kind_rev
-from .._digitization.base import _dig_kind_ints
-from .._digitization._utils import _read_dig_fif
 
 b = bytes  # alias
 
@@ -706,8 +703,7 @@ def write_fiducials(fname, pts, coord_frame=FIFF.FIFFV_COORD_UNKNOWN,
         mne.io.constants.FIFF.FIFFV_COORD_...).
     %(verbose)s
     """
-    from .._digitization._utils import write_dig as ff
-    ff(fname, pts, coord_frame)
+    _dig_write_dig(fname, pts, coord_frame)
 
 
 def write_dig(fname, pts, coord_frame=None):
@@ -725,8 +721,7 @@ def write_dig(fname, pts, coord_frame=None):
         here. Can be None (default) if the points could have varying
         coordinate frames.
     """
-    from .._digitization._utils import write_dig as ff
-    return ff(fname, pts, coord_frame=None)
+    return _dig_write_dig(fname, pts, coord_frame=None)
 
 
 @verbose
